+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- hnefatafl-server-full: leave_game.
 - hnefatafl-server-full: save the general chat texts when restarting the server.
 - hnefatafl-server-full: persist request_draw.